requirements.txt
torch
openai-whisper
whisperx
python-dotenv
flask
python-docx
mistralai
